requirements-1.txt
opencv-python
imutils
numpy
Pillow
